requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
anthropic>=0.19.1
openai==1.3.9
python-multipart==0.0.6
gunicorn==21.2.0
httpx==0.25.0
uuid==1.30